+++ wallet/walletrpc.c
@@ -4,6 +4,7 @@
 #include <ccan/tal/str/str.h>
 #include <common/bech32.h>
 #include <common/json_command.h>
+#include <common/json_helpers.h>
 #include <common/jsonrpc_errors.h>
 #include <common/key_derive.h>
 #include <common/param.h>
@@ -45,33 +46,29 @@ struct withdrawal {
  */
 static void wallet_withdrawal_broadcast(struct bitcoind *bitcoind UNUSED,
 					int exitstatus, const char *msg,
-					struct withdrawal *withdraw)
+					struct unreleased_tx *utx)
 {
-	struct command *cmd = withdraw->cmd;
-	struct lightningd *ld = withdraw->cmd->ld;
+	struct command *cmd = utx->wtx->cmd;
+	struct lightningd *ld = cmd->ld;
 	struct amount_sat change = AMOUNT_SAT(0);
 
+	/* FIXME: This won't be necessary once we use ccan/json_out! */
 	/* Massage output into shape so it doesn't kill the JSON serialization */
 	char *output = tal_strjoin(cmd, tal_strsplit(cmd, msg, "\n", STR_NO_EMPTY), " ", STR_NO_TRAIL);
 	if (exitstatus == 0) {
 		/* Mark used outputs as spent */
-		wallet_confirm_utxos(ld->wallet, withdraw->wtx->utxos);
-
-		/* Parse the tx and extract the change output. We
-		 * generated the hex tx, so this should always work */
-		struct bitcoin_tx *tx = bitcoin_tx_from_hex(withdraw, withdraw->hextx, strlen(withdraw->hextx));
-		assert(tx != NULL);
+		wallet_confirm_utxos(ld->wallet, utx->wtx->utxos);
 
 		/* Extract the change output and add it to the DB */
-		wallet_extract_owned_outputs(ld->wallet, tx, NULL, &change);
+		wallet_extract_owned_outputs(ld->wallet, utx->tx, NULL, &change);
 
 		/* Note normally, change_satoshi == withdraw->wtx->change, but
 		 * not if we're actually making a payment to ourselves! */
-		assert(amount_sat_greater_eq(change, withdraw->wtx->change));
+		assert(amount_sat_greater_eq(change, utx->wtx->change));
 
 		struct json_stream *response = json_stream_success(cmd);
 		json_object_start(response, NULL);
-		json_add_string(response, "tx", withdraw->hextx);
+		json_add_tx(response, "tx", utx->tx);
 		json_add_string(response, "txid", output);
 		json_object_end(response);
 		was_pending(command_success(cmd, response));
@@ -106,6 +103,232 @@ static struct command_result *param_bitcoin_address(struct command *cmd,
 	abort();
 }
 
+/* Signs the tx, broadcasts it: broadcast calls wallet_withdrawal_broadcast */
+static struct command_result *broadcast_and_wait(struct command *cmd,
+						 struct unreleased_tx *utx)
+{
+	struct bitcoin_tx *signed_tx;
+	struct bitcoin_txid signed_txid;
+
+	/* FIXME: hsm will sign almost anything, but it should really
+	 * fail cleanly (not abort!) and let us report the error here. */
+	u8 *msg = towire_hsm_sign_withdrawal(cmd,
+					     utx->wtx->amount,
+					     utx->wtx->change,
+					     utx->wtx->change_key_index,
+					     utx->destination,
+					     utx->wtx->utxos);
+
+	if (!wire_sync_write(cmd->ld->hsm_fd, take(msg)))
+		fatal("Could not write sign_withdrawal to HSM: %s",
+		      strerror(errno));
+
+	msg = wire_sync_read(cmd, cmd->ld->hsm_fd);
+
+	if (!fromwire_hsm_sign_withdrawal_reply(utx, msg, &signed_tx))
+		fatal("HSM gave bad sign_withdrawal_reply %s",
+		      tal_hex(tmpctx, msg));
+
+	/* Sanity check */
+	bitcoin_txid(signed_tx, &signed_txid);
+	if (!bitcoin_txid_eq(&signed_txid, &utx->txid))
+		fatal("HSM changed txid: unsigned %s, signed %s",
+		      tal_hex(tmpctx, linearize_tx(tmpctx, utx->tx)),
+		      tal_hex(tmpctx, linearize_tx(tmpctx, signed_tx)));
+
+	/* Replace unsigned tx by signed tx. */
+	tal_free(utx->tx);
+	utx->tx = signed_tx;
+
+	/* Now broadcast the transaction */
+	bitcoind_sendrawtx(cmd->ld->topology->bitcoind,
+			   tal_hex(tmpctx, linearize_tx(tmpctx, signed_tx)),
+			   wallet_withdrawal_broadcast, utx);
+
+	return command_still_pending(cmd);
+}
+
+/* Common code for withdraw and txprepare.
+ *
+ * Returns NULL on success, and fills in wtx, destination and
+ * maybe changekey (owned by cmd).  Otherwise, cmd has failed, so don't
+ * access it! (It's been freed). */
+static struct command_result *json_prepare_tx(struct command *cmd,
+					      const char *buffer,
+					      const jsmntok_t *params,
+					      struct unreleased_tx **utx)
+{
+	u32 *feerate_per_kw;
+	struct command_result *res;
+	u32 *minconf, maxheight;
+	struct pubkey *changekey;
+
+	*utx = tal(cmd, struct unreleased_tx);
+	(*utx)->wtx = tal(*utx, struct wallet_tx);
+	wtx_init(cmd, (*utx)->wtx, AMOUNT_SAT(-1ULL));
+
+	if (!param(cmd, buffer, params,
+		   p_req("destination", param_bitcoin_address,
+			 &(*utx)->destination),
+		   p_req("satoshi", param_wtx, (*utx)->wtx),
+		   p_opt("feerate", param_feerate, &feerate_per_kw),
+		   p_opt_def("minconf", param_number, &minconf, 1),
+		   NULL))
+		return command_param_failed();
+
+	/* Destination is owned by cmd: change that to be owned by utx. */
+	tal_steal(*utx, (*utx)->destination);
+
+	if (!feerate_per_kw) {
+		res = param_feerate_estimate(cmd, &feerate_per_kw,
+					     FEERATE_NORMAL);
+		if (res)
+			return res;
+	}
+
+	maxheight = minconf_to_maxheight(*minconf, cmd->ld);
+	res = wtx_select_utxos((*utx)->wtx, *feerate_per_kw,
+			       tal_count((*utx)->destination), maxheight);
+	if (res)
+		return res;
+
+	if (!amount_sat_eq((*utx)->wtx->change, AMOUNT_SAT(0))) {
+		changekey = tal(tmpctx, struct pubkey);
+		if (!bip32_pubkey(cmd->ld->wallet->bip32_base, changekey,
+				  (*utx)->wtx->change_key_index))
+			return command_fail(cmd, LIGHTNINGD, "Keys generation failure");
+	}
+
+	(*utx)->tx = withdraw_tx(*utx, (*utx)->wtx->utxos,
+				 (*utx)->destination, (*utx)->wtx->amount,
+				 changekey, (*utx)->wtx->change,
+				 cmd->ld->wallet->bip32_base,
+				 &(*utx)->change_outnum);
+	bitcoin_txid((*utx)->tx, &(*utx)->txid);
+
+	return NULL;
+}
+
+static struct command_result *json_txprepare(struct command *cmd,
+					     const char *buffer,
+					     const jsmntok_t *obj UNNEEDED,
+					     const jsmntok_t *params)
+{
+	struct unreleased_tx *utx;
+	struct command_result *res;
+	struct json_stream *response;
+
+	res = json_prepare_tx(cmd, buffer, params, &utx);
+	if (res)
+		return res;
+
+	/* utx will persist past this command. */
+	tal_steal(cmd->ld->wallet, utx);
+	add_unreleased_tx(cmd->ld->wallet, utx);
+
+	response = json_stream_success(cmd);
+	json_object_start(response, NULL);
+	json_add_tx(response, "unsigned_tx", utx->tx);
+	json_add_txid(response, "txid", &utx->txid);
+	json_object_end(response);
+	return command_success(cmd, response);
+}
+static const struct json_command txprepare_command = {
+	"txprepare",
+	"bitcoin",
+	json_txprepare,
+	"Create a transaction, with option to spend in future (either txsend and txdiscard)",
+	false
+};
+AUTODATA(json_command, &txprepare_command);
+
+static struct command_result *param_unreleased_txid(struct command *cmd,
+						    const char *name,
+						    const char *buffer,
+						    const jsmntok_t *tok,
+						    struct unreleased_tx **utx)
+{
+	struct command_result *res;
+	struct bitcoin_txid *txid;
+
+	res = param_txid(cmd, name, buffer, tok, &txid);
+	if (res)
+		return res;
+
+	*utx = find_unreleased_tx(cmd->ld->wallet, txid);
+	if (!*utx)
+		return command_fail(cmd, LIGHTNINGD,
+				    "%s not an unreleased txid",
+				    type_to_string(cmd, struct bitcoin_txid,
+						   txid));
+	tal_free(txid);
+	return NULL;
+}
+
+static struct command_result *json_txsend(struct command *cmd,
+					  const char *buffer,
+					  const jsmntok_t *obj UNNEEDED,
+					  const jsmntok_t *params)
+{
+	struct unreleased_tx *utx;
+
+	if (!param(cmd, buffer, params,
+		   p_req("txid", param_unreleased_txid, &utx),
+		   NULL))
+		return command_param_failed();
+
+	/* We delete from list now, and this command owns it. */
+	remove_unreleased_tx(utx);
+	tal_steal(cmd, utx);
+
+	/* We're the owning cmd now. */
+	utx->wtx->cmd = cmd;
+
+	return broadcast_and_wait(cmd, utx);
+}
+
+static const struct json_command txsend_command = {
+	"txsend",
+	"bitcoin",
+	json_txsend,
+	"Sign and broadcast a transaction created by txprepare",
+	false
+};
+AUTODATA(json_command, &txsend_command);
+
+static struct command_result *json_txdiscard(struct command *cmd,
+					     const char *buffer,
+					     const jsmntok_t *obj UNNEEDED,
+					     const jsmntok_t *params)
+{
+	struct unreleased_tx *utx;
+	struct json_stream *response;
+
+	if (!param(cmd, buffer, params,
+		   p_req("txid", param_unreleased_txid, &utx),
+		   NULL))
+		return command_param_failed();
+
+	/* Free utx with this command */
+	tal_steal(cmd, utx);
+
+	response = json_stream_success(cmd);
+	json_object_start(response, NULL);
+	json_add_tx(response, "unsigned_tx", utx->tx);
+	json_add_txid(response, "txid", &utx->txid);
+	json_object_end(response);
+	return command_success(cmd, response);
+}
+
+static const struct json_command txdiscard_command = {
+	"txdiscard",
+	"bitcoin",
+	json_txdiscard,
+	"Abandon a transaction created by txprepare",
+	false
+};
+AUTODATA(json_command, &txdiscard_command);
+
 /**
  * json_withdraw - Entrypoint for the withdrawal flow
  *
@@ -118,68 +341,14 @@ static struct command_result *json_withdraw(struct command *cmd,
 					    const jsmntok_t *obj UNNEEDED,
 					    const jsmntok_t *params)
 {
-	struct withdrawal *withdraw = tal(cmd, struct withdrawal);
-	u32 *feerate_per_kw;
-	struct bitcoin_tx *tx;
-	struct pubkey changekey;
+	struct unreleased_tx *utx;
 	struct command_result *res;
-	u32 *minconf, maxheight;
 
-	withdraw->cmd = cmd;
-	withdraw->wtx = tal(withdraw, struct wallet_tx);
-	wtx_init(cmd, withdraw->wtx, AMOUNT_SAT(-1ULL));
-
-	if (!param(cmd, buffer, params,
-		   p_req("destination", param_bitcoin_address,
-			 (const u8 **)&withdraw->destination),
-		   p_req("satoshi", param_wtx, withdraw->wtx),
-		   p_opt("feerate", param_feerate, &feerate_per_kw),
-		   p_opt_def("minconf", param_number, &minconf, 1),
-		   NULL))
-		return command_param_failed();
-
-	if (!feerate_per_kw) {
-		res = param_feerate_estimate(cmd, &feerate_per_kw,
-					     FEERATE_NORMAL);
-		if (res)
-			return res;
-	}
-
-	maxheight = minconf_to_maxheight(*minconf, cmd->ld);
-	res = wtx_select_utxos(withdraw->wtx, *feerate_per_kw,
-			       tal_count(withdraw->destination), maxheight);
+	res = json_prepare_tx(cmd, buffer, params, &utx);
 	if (res)
 		return res;
 
-	if (!bip32_pubkey(cmd->ld->wallet->bip32_base, &changekey,
-			  withdraw->wtx->change_key_index)) {
-		return command_fail(cmd, LIGHTNINGD, "Keys generation failure");
-	}
-	txfilter_add_scriptpubkey(cmd->ld->owned_txfilter,
-				  scriptpubkey_p2wpkh(tmpctx, &changekey));
-
-	u8 *msg = towire_hsm_sign_withdrawal(cmd,
-					     withdraw->wtx->amount,
-					     withdraw->wtx->change,
-					     withdraw->wtx->change_key_index,
-					     withdraw->destination,
-					     withdraw->wtx->utxos);
-
-	if (!wire_sync_write(cmd->ld->hsm_fd, take(msg)))
-		fatal("Could not write sign_withdrawal to HSM: %s",
-		      strerror(errno));
-
-	msg = wire_sync_read(cmd, cmd->ld->hsm_fd);
-
-	if (!fromwire_hsm_sign_withdrawal_reply(msg, msg, &tx))
-		fatal("HSM gave bad sign_withdrawal_reply %s",
-		      tal_hex(withdraw, msg));
-
-	/* Now broadcast the transaction */
-	withdraw->hextx = tal_hex(withdraw, linearize_tx(cmd, tx));
-	bitcoind_sendrawtx(cmd->ld->topology->bitcoind, withdraw->hextx,
-			   wallet_withdrawal_broadcast, withdraw);
-	return command_still_pending(cmd);
+	return broadcast_and_wait(cmd, utx);
 }
 
 static const struct json_command withdraw_command = {